refactor: define organization for all unit test
